# documentation/openapi.yaml
openapi: 3.0.3
info:
  title: Club Accounting App API
  version: 1.0.0
  description: >
    REST API for a web based accounting application.
servers:
  - url: https://localhost:8443
    description: Local development environment (HTTPS)
  - url: http://localhost:8080
    description: Local development environment (HTTP)

paths:
  /projects:
    post:
      tags:
        - Project
      summary: Create a new accounting project
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "201":
          description: Project successfully created
        "400":
          description: Invalid input

  /login:
    post:
      tags:
        - Project
      summary: Log in to an existing accounting project
      operationId: loginProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        "401":
          description: Invalid credentials

  /accounts:
    get:
      tags:
        - Account
      summary: Get all accounts of this accounting project
      operationId: getAccounts
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    put:
      tags:
        - Account
      summary: Save adjusted accounts
      operationId: updateAccounts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                accounts:
                  type: array
                  items:
                    $ref: "#/components/schemas/AccountUpdate"
      responses:
        "204":
          description: |
            Changes processed successfully (no content)
            
            Behavior:
            - If an item contains a new `number` with a non-empty `name`, the account is created.
            - If an item references an existing `number` and contains a non-empty `name`, the account's name is updated.
            - If an item references an existing `number` and contains `name: null`, the account is deleted.
            - If an item contains a non-existing `number` and `name` is null or blank, it is ignored (no-op).
        "409":
          description: Conflict – account cannot be deleted due to existing references (e.g., bookings)

  /bookings:
    get:
      tags:
        - Booking
      summary: Get all bookings of this accounting project
      operationId: getBookings
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    put:
      tags:
        - Booking
      summary: Save adjusted bookings
      operationId: updateBookings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                entries:
                  type: array
                  items:
                    $ref: "#/components/schemas/BookingUpdate"
      responses:
        "204":
          description: |
            Changes processed successfully (no content)

            Behavior:
            - Create: When an item has `id: null` and all fields `date`, `text`, `debit`, `credit`, `amount` are present with non-null values, a booking is created.
            - Update: When an item has an existing `id`, any provided non-null field updates that attribute. Omitted fields remain unchanged.
            - Delete via PUT: When an item has an existing `id` and contains only fields explicitly set to `null` (and no non-null updates), the booking is deleted.
            - Debit/Credit reference accounts by their `number` (account_number), not by internal database id.
            - Unknown ids: Items that reference a non-existing `id` are ignored (no-op).
        "400":
          description: Bad Request – invalid account references or missing fields for create
        "401":
          description: Unauthorized
        "403":
          description: Forbidden – booking does not belong to the project

  /bookings/{id}:
    delete:
      tags:
        - Booking
      summary: Delete a booking by id
      operationId: deleteBooking
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted successfully
        "401":
          description: Unauthorized
        "403":
          description: Forbidden – booking does not belong to the project
        "404":
          description: Booking not found

components:
  schemas:
    LoginRequest:
      type: object
      required: [project_name, password]
      properties:
        project_name:
          type: string
          example: Hobby_Horsing_Horgen
        password:
          type: string
          format: password
          example: geheim123

    Account:
      type: object
      required: [number, name]
      properties:
        number:
          type: integer
          example: 1000
        name:
          type: string
          example: Kasse

    AccountUpdate:
      type: object
      required: [number]
      properties:
        number:
          type: integer
          example: 1020
        name:
          type: string
          nullable: true
          example: Bank
          description: |
            The account name. When omitted, the name remains unchanged. When set
            to a non-empty value, the name is updated/used on create. When explicitly
            set to null, the account will be deleted if it exists.

    Booking:
      type: object
      required: [number, date, text, debit, credit, amount]
      properties:
        number:
          type: integer
          example: 15
        date:
          type: string
          format: date
          example: 2025-03-15
        text:
          type: string
          example: Bargeld abgehoben
        debit:
          type: integer
          example: 1000
        credit:
          type: integer
          example: 1020
        amount:
          type: number
          format: float
          example: 200.00

    BookingUpdate:
      type: object
      properties:
        id:
          type: integer
          example: 17
        date:
          type: string
          format: date
          nullable: true
        text:
          type: string
          nullable: true
        debit:
          type: integer
          nullable: true
        credit:
          type: integer
          nullable: true
        amount:
          type: number
          format: float
          nullable: true